Coalesced all module control configutation into a single file, testconfig.yml
This is a dict with possible keys of 'exception', 'preload' and 'skip'.

The modules in preload are loaded before the walk_packages as there is some
order of operation issue with the tests that isn't evident in every day use.

The modules in skip are not loaded when packages are walked, as they cause
errors that do gracefully exit with clean exceptions.

The modules in exception are allowed to be loaded, fail, and raise an exception,
but these are monitored and any untripped exception is reported. In this way
they can be removed when no longer required.